flask/requirements.txt
alembic==1.2.1
aniso8601==8.0.0
certifi==2019.11.28
Click==7.0
cloudinary==1.18.2
Faker==2.0.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mock==3.0.5
Pillow==6.2.1
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
reportlab==3.5.32
six==1.12.0
SQLAlchemy==1.3.10
text-unidecode==1.3
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
